Organize image generation
