+++ src/lib/images/copyFileToPublic.ts
@@ -1,0 +1,20 @@
+import { copyFileSync, existsSync, mkdirSync } from 'fs'
+import path from 'path'
+
+export const copyFileToPublic = (
+  fileName: string,
+  sourceFile: string,
+  targetFolder: string
+) => {
+  if (!existsSync(targetFolder)) {
+    mkdirSync(targetFolder, { recursive: true })
+  }
+
+  if (existsSync(path.join(targetFolder, fileName))) {
+    return
+  }
+
+  // eslint-disable-next-line no-console
+  console.log(`Copying ${fileName} to the public folder`)
+  copyFileSync(sourceFile, path.join(targetFolder, fileName))
+}
